refactor: centralize file preprocessing and add document extractors
