+++ requirements.txt
@@ -8,6 +8,8 @@ psycopg[binary]
 psycopg2==2.9.12
 python-docx==1.2.0
 python-dotenv
+python-pptx==1.0.2
+xlrd==2.0.1
 pywin32==311; platform_system == "Windows"
 pyhwp==0.1b15
 requests==2.33.0
