+++ scripts/extractors/__init__.py
@@ -1,0 +1,6 @@
+from __future__ import annotations
+
+from scripts.extractors.common import extract_blocks
+from scripts.extractors.pptx import extract_pptx_blocks
+
+__all__ = ["extract_blocks", "extract_pptx_blocks"]
